fix(omo-codex): validate cached plugin bin links
